LayoutTests/imported/w3c/web-platform-tests/fetch/metadata/unload.https.sub-expected.txt



PASS Fetch from an unload handler
FAIL undefined: sec-fetch-dest assert_equals: expected "empty" but got ""
FAIL undefined: sec-fetch-mode assert_equals: expected "no-cors" but got ""
FAIL undefined: sec-fetch-site assert_equals: expected "same-origin" but got ""
PASS undefined: sec-fetch-user
